backend/requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
pymongo==4.6.0
python-dotenv==1.0.0
